bot/requirements.txt
python-telegram-bot
pymongo
gspread
oauth2client